# apollo-openapi.yaml
openapi: 3.0.1
info:
  title: Apollo OpenAPI
  description: |
    <p>Apollo配置中心OpenAPI接口文档</p>

    <h2>认证方式</h2>
    <p>所有 API 接口都需要通过 <code>Authorization</code> header 进行身份验证。</p>

    <h3>获取 Token 的方式：</h3>
    <ol>
      <li><b>Portal 管理界面获取</b>：登录 Portal → 管理员工具 → 开放平台授权管理 → 创建第三方应用，获取 Token。</li>
      <li><b>Token 格式</b>：<code>Authorization: token_value</code></li>
      <li><b>Token 权限</b>：按应用/环境/命名空间授予，建议不同用途分别创建。</li>
    </ol>

    <h3>使用示例</h3>
    <pre><code class="language-bash">curl -X GET "http://localhost:8070/openapi/v1/apps" \
    -H "Authorization: your_token_here"</code></pre>

  version: 1.0.0
security:
  - ApiKeyAuth: []
...
paths:
  /openapi/v1/apps:
    post:
      summary: 创建应用 (original openapi) # which means this belongs to initial openapi endpoints and should not be modified this time
      operationId: createApp
      deprecated: false
      description: POST /openapi/v1/apps
      tags:
        - App Management
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenCreateAppDTO' # this is for avoiding openapi-generator of generating a new schema(OpenAppDTO with restriction)
        required: true
      responses:
        '200':
          description: '应用创建成功'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAppDTO'
          headers: {}
        '400':
          description: '请求参数错误'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse' # This is to avoid openapi-generator generating completely different error responses for each error
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
    get:
      summary: 查找应用 (original openapi)
      operationId: findApps
      deprecated: false
      description: POST /openapi/v1/apps
      tags:
        - App Management
      parameters:
        - name: appIds
          in: query
          description: 查找应用
          required: false
          example: '123'
          schema:
            type: string
      responses:
        '200':
          description: 成功获取应用列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpenAppDTO'
              example:
                - dataChangeCreatedBy: 'apollo'
                  dataChangeLastModifiedBy: 'apollo'
                  dataChangeCreatedTime: '2024-01-15T10:30:00.000Z'
                  dataChangeLastModifiedTime: '2024-01-15T10:30:00.000Z'
                  name: 'Sample Application'
                  appId: 'sample-app'
                  orgId: 'default'
                  orgName: 'Default Organization'
                  ownerName: 'Apollo Admin'
                  ownerEmail: 'admin@apollo.com'
                - dataChangeCreatedBy: 'apollo'
                  dataChangeLastModifiedBy: 'apollo'
                  dataChangeCreatedTime: '2024-01-10T14:20:00.000Z'
                  dataChangeLastModifiedTime: '2024-01-12T16:45:00.000Z'
                  name: 'User Service'
                  appId: 'user-service'
                  orgId: 'microservices'
                  orgName: 'Microservices Team'
                  ownerName: 'Dev Team'
                  ownerEmail: 'dev@company.com'
          headers: {}
        '401':
          description: '未授权访问'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
  /openapi/v1/apps/authorized:
    get:
      summary: 获取当前Consumer授权的应用列表 (original openapi)
      operationId: findAppsAuthorized
      deprecated: false
      description: GET /openapi/v1/apps/authorized
      tags:
        - App Management
      responses:
        '200':
          description: 成功获取授权应用列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpenAppDTO'
              example:
                - dataChangeCreatedBy: 'apollo'
                  dataChangeLastModifiedBy: 'apollo'
                  dataChangeCreatedTime: '2024-01-15T10:30:00.000Z'
                  dataChangeLastModifiedTime: '2024-01-15T10:30:00.000Z'
                  name: 'Sample Application'
                  appId: 'sample-app'
                  orgId: 'default'
                  orgName: 'Default Organization'
                  ownerName: 'Apollo Admin'
                  ownerEmail: 'admin@apollo.com'
                - dataChangeCreatedBy: 'apollo'
                  dataChangeLastModifiedBy: 'apollo'
                  dataChangeCreatedTime: '2024-01-10T14:20:00.000Z'
                  dataChangeLastModifiedTime: '2024-01-12T16:45:00.000Z'
                  name: 'User Service'
                  appId: 'user-service'
                  orgId: 'microservices'
                  orgName: 'Microservices Team'
                  ownerName: 'Dev Team'
                  ownerEmail: 'dev@company.com'
          headers: { }
        '401':
          description: '未授权访问'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
  /openapi/v1/apps/{appId}/envclusters:
    get:
      summary: 获取应用的环境集群信息 (original openapi)
      operationId: getEnvClusters
      deprecated: false
      description: GET /openapi/v1/apps/{appId}/envclusters
      tags:
        - App Management
      parameters:
        - name: appId
          in: path
          description: 应用ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: '成功获取应用环境集群信息'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpenEnvClusterDTO'
              example:
                - env: 'DEV'
                  clusters:
                    - 'default'
                    - 'test-cluster'
                - env: 'UAT'
                  clusters:
                    - 'default'
                - env: 'PRO'
                  clusters:
                    - 'default'
                    - 'backup'
          headers: {}
        '404':
          description: '应用不存在'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
  /openapi/v1/apps/{appId}:
    get:
      summary: 获取单个应用信息(new added) # which means this is a new added openapi endpoint
      operationId: getApp
      deprecated: false
      description: GET /openapi/v1/apps/{appId}
      tags:
        - App Management
      parameters:
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: '成功获取应用信息'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAppDTO'
              example:
                dataChangeCreatedBy: 'apollo'
                dataChangeLastModifiedBy: 'dev-user'
                dataChangeCreatedTime: '2024-01-15T10:30:00.000Z'
                dataChangeLastModifiedTime: '2024-01-20T15:45:00.000Z'
                name: 'Sample Application'
                appId: 'sample-app'
                orgId: 'default'
                orgName: 'Default Organization'
                ownerName: 'Apollo Admin'
                ownerEmail: 'admin@apollo.com'
          headers: {}
        '404':
          description: '应用不存在'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
    put:
      summary: 更新应用(new added)
      operationId: updateApp
      deprecated: false
      description: PUT /openapi/v1/apps/{appId}
      tags:
        - App Management
      parameters:
        - name: appId
          in: path
          description: '应用ID'
          required: true
          schema:
            type: string
        - name: operator
          in: query
          description: '操作人用户名'
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenAppDTO'
        required: true
      responses:
        '200':
          description: '应用更新成功'
        '400':
          description: '请求参数错误'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        '403':
          description: '权限不足'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
    delete:
      summary: 删除应用(new added)
      operationId: deleteApp
      deprecated: false
      description: DELETE /openapi/v1/apps/{appId}
      tags:
        - App Management
      parameters:
        - name: appId
          in: path
          description: '应用ID'
          required: true
          schema:
            type: string
        - name: operator
          in: query
          description: '操作人用户名'
          required: false
          schema:
            type: string
      responses:
        '200':
          description: '应用删除成功'
        '403':
          description: '权限不足，需要超级管理员权限'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        '404':
          description: '应用不存在'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
          headers: {}
  /openapi/v1/apps/by-self:
    get:
      summary: 获取当前Consumer/User的应用列表（分页）(new added)
      operationId: getAppsBySelf
      deprecated: false
      description: GET /openapi/v1/apps/by-self
      tags:
        - App Management
      parameters:
        - name: page
          in: query
          description: '页数'
          required: true
          schema:
            type: integer
            minimum: 0
        - name: size
          in: query
          description: '页大小'
          required: true
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: '成功获取Consumer的应用列表'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpenAppDTO'
              example:
                - dataChangeCreatedBy: 'apollo'
                  dataChangeLastModifiedBy: 'dev-user'
                  dataChangeCreatedTime: '2024-01-15T10:30:00.000Z'
                  dataChangeLastModifiedTime: '2024-01-20T16:45:00.000Z'
                  name: 'User Service'
                  appId: 'user-service'
                  orgId: 'microservices'
                  orgName: 'Microservices Team'
                  ownerName: 'Dev Team'
                  ownerEmail: 'dev@company.com'
                - dataChangeCreatedBy: 'apollo'
                  dataChangeLastModifiedBy: 'ops-user'
                  dataChangeCreatedTime: '2024-01-10T09:15:00.000Z'
                  dataChangeLastModifiedTime: '2024-01-18T14:30:00.000Z'
                  name: 'Order Service'
                  appId: 'order-service'
                  orgId: 'business'
                  orgName: 'Business Team'
                  ownerName: 'Business Team'
                  ownerEmail: 'business@company.com'
          headers: {}
        '401':
          description: '未授权访问'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
  /openapi/v1/apps/{appId}/env-cluster-info:
    get:
      summary: 获取应用环境集群详情(new added)
      operationId: getEnvClusterInfo
      deprecated: false
      description: /openapi/v1/apps/{appId}/env-cluster-info
      tags:
        - App Management
      parameters:
        - name: appId
          in: path
          description: '应用ID'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: '成功获取应用环境集群详情'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpenEnvClusterInfo'
              example:
                - env: 'DEV'
                  code: '200'
                  clusters:
                    - dataChangeCreatedBy: 'apollo'
                      dataChangeLastModifiedBy: 'apollo'
                      dataChangeCreatedByDisplayName: 'apollo'
                      dataChangeLastModifiedByDisplayName: 'apollo'
                      dataChangeCreatedTime: '2023-01-01T00:00:00.000Z'
                      dataChangeLastModifiedTime: '2023-01-01T00:00:00.000Z'
                      id: 1
                      name: 'default'
                      appId: 'sample-app'
                      parentClusterId: 0
                      comment: 'Default cluster'
                  message: 'load env: DEV cluster error'
          headers: {}
  /openapi/v1/apps/envs/{env}:
    post:
      summary: 在指定环境创建应用(new added)
      operationId: createAppInEnv
      deprecated: false
      description: POST /openapi/v1/apps/envs/{env}
      tags:
        - App Management
      parameters:
        - name: env
          in: path
          description: 环境标识，例如 DEV、FAT、UAT、PROD
          required: true
          schema:
            type: string
        - name: operator
          in: query
          description: '操作人用户名'
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenAppDTO'
        required: true
      responses:
        '200':
          description: '应用在指定环境创建成功'
        '400':
          description: '请求参数错误'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        '403':
          description: '权限不足'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
          headers: {}
  /openapi/v1/apps/{appId}/miss-envs:
    get:
      summary: 查找缺失的环境(new added)
      operationId: findMissEnvs
      deprecated: false
      description: GET /openapi/v1/apps/{appId}/miss-envs
      tags:
        - App Management
      parameters:
        - name: appId
          in: path
          description: '应用ID'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功获取应用缺失的环境列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpenMissEnvDTO'
              example:
                - code: "200"
                  message: "test_NOT_FOUND"
                - code: "500"
                  message: "load appId:xx from env test_INTERNAL_ERROR error. test ERROR MESSAGE"
        '404':
          description: '应用不存在'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Apollo OpenAPI访问令牌
        **获取方式：**
        1. 登录Apollo Portal管理界面
        2. 进入"管理员工具" -> "开放平台授权管理"
        3. 创建第三方应用，获取Token
        **使用格式：**
        ```
        Authorization: your_token_here
        ```
        **权限说明：**
        - Token具有对应应用的读写权限
        - 不同Token可能有不同的环境和命名空间权限
        - 建议为不同用途创建不同的Token
  schemas:
    OpenAppDTO:
      type: object
      properties:
        dataChangeCreatedBy:
          type: string
          description: 数据创建者用户名，记录是谁创建了这个应用
        dataChangeLastModifiedBy:
          type: string
          description: 数据最后修改者用户名，记录最后一次修改应用信息的用户
        dataChangeCreatedTime:
          type: string
          description: 数据创建时间，ISO 8601格式的时间戳
          example: "2025-09-29T12:34:56Z"
        dataChangeLastModifiedTime:
          type: string
          description: 数据最后修改时间，ISO 8601格式的时间戳
          example: "2025-09-29T12:34:56Z"
        name:
          type: string
          description: 应用名称，用于显示的友好名称
        appId:
          type: string
          description: 应用唯一标识符，全局唯一的应用ID
        orgId:
          type: string
          description: 组织ID，应用所属组织的唯一标识
        orgName:
          type: string
          description: 组织名称，应用所属组织的显示名称
        ownerName:
          type: string
          description: 应用负责人姓名，应用的主要负责人
        ownerEmail:
          type: string
          description: 应用负责人邮箱地址，用于接收应用相关通知
        ownerDisplayName:
          type: string
      description: Apollo应用信息数据传输对象，包含应用的基本信息和元数据
    OpenCreateAppDTO:
      type: object
      properties:
        assignAppRoleToSelf:
          type: boolean
          description: 是否将应用角色分配给当前用户。当设置为true时，当前token用户将获得该应用的完全操作权限
        admins:
          type: array
          items:
            type: string
          uniqueItems: true
          description: 应用管理员用户名列表。应用负责人默认具有项目管理员权限。管理员可以创建命名空间、集群并分配用户权限
        app:
          $ref: '#/components/schemas/OpenAppDTO'
      description: 创建Apollo应用的请求数据传输对象，包含创建应用所需的所有信息
...
    OpenClusterDTO:
      type: object
      properties:
        dataChangeCreatedBy:
          type: string
          description: 集群创建者用户名，记录是谁创建了这个集群
        dataChangeLastModifiedBy:
          type: string
          description: 集群最后修改者用户名，记录最后一次修改集群信息的用户
        dataChangeCreatedTime:
          type: string
          description: 集群创建时间，ISO 8601格式的时间戳
          example: "2025-09-29T12:34:56Z"
        dataChangeLastModifiedTime:
          type: string
          description: 集群最后修改时间，ISO 8601格式的时间戳
          example: "2025-09-29T12:34:56Z"
        name:
          type: string
          description: 集群名称，在同一应用和环境下唯一标识一个集群
        appId:
          type: string
          description: 所属应用的唯一标识符
        id:
          type: integer
          description: 集群的唯一标识符
          format: int64
        parentClusterId:
          type: integer
          description: 父集群的ID
          format: int64
        comment:
          type: string
          description: 集群的备注说明
      description: Apollo集群信息数据传输对象，表示应用在特定环境下的集群配置
...
    OpenEnvClusterDTO:
      type: object
      properties:
        env:
          type: string
          description: '环境标识，例如 DEV、FAT、UAT、PROD'
        clusters:
          type: array
          items:
            type: string
          description: 该环境下的集群名称列表
    OpenEnvClusterInfo:
      type: object
      properties:
        code:
          type: integer
          description: 查询状态码
        message:
          type: string
          description: '相关信息'
        env:
          type: string
          description: '环境标识'
        clusters:
          type: array
          items:
            $ref: '#/components/schemas/OpenClusterDTO'
          description: '集群信息列表'
    OpenMissEnvDTO:
      type: object
      properties:
        code:
          type: integer
          description: 查询状态码
        message:
          type: string
...
    ExceptionResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP 状态码
        message:
          type: string
          description: 错误信息
        timestamp:
          type: string
          format: date-time
          description: 时间戳
        exception:
          type: string
          description: 错误类名
      example:
        status: 404
        message: "Resource not found"
        timestamp: "2025-09-29T12:34:56Z"
        exception: "NotFoundException"
...
servers: []